Moved services page to sperate file
